Simplify entrypoint code
